Changing weights on later DI solves
For the new kMS/DDFacet a bug was fixed so that the DI weights are actually applied. Turns out this isnt good. 

Here we tell kMS not to update the weights on the DI calibrations that come after DD calibrations. Hence DD weights (rather than DI ones) will be used for the subsequent imaging.

The very first DI calibration still updates the weights because we no not have any DD weights yet.